AdamBot has its own object file, cleaned up and separated out cog loading
